// src/types/vehicleRequest.ts
/**
 * Generated by orval v8.5.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { VehicleRequestType } from './vehicleRequestType';

export interface VehicleRequest {
  licensePlate?: string;
  type?: VehicleRequestType;
  maxWeight?: number;
  maxVolume?: number;
  shipperId?: string;
}